AAP: Replace memsrc6 with memsrc10, update instrs
Memsrc10 is a ten bit signed offset for load and store instructions,
replacing memsrc6. It reuses the opcodes bits of the second instruction
word to encode a longer immediate.

A new fixup and relocation has been added for the 10 bit offset field.
The field encodes in the same bits as the ABS10 relocation, the only
difference is that it may be signed.