AAP: Remove short fixups+relocs, add bias to shift instructions
The fixups and relocations have been rearranged. The ABS*_SHORT fixups
have been removed, as have their accompanying relocations. There is no
circumstance where these should be emitted.

Shift operands now have a bias applied to them, so that values are
encoded with a bias of -1. A zero shift amount is no longer valid. This
is inline with the latest revision of the architecture.

A new fixup+relocation has been added for shift immediates.