AAP: Don't save callee saved regs used for return
If a register is used for the return, it does not need to be callee
saved, as the value will be clobbered anyway. The return register is a
def at the call site, so it will be caller saved instead.

This adds a new hook in MachineRegisterInfo which allows a target to
specify registers which are live out of a function. When determining the
callee saved registers, these live-out registers are ignored.

Without this hook Prologue/Epilogue inserter would introduce a spill and
reload for a return register, and this would cause the return value to
be overwritten.