Implementation of register resurrection optimization
This implements an optimization which allows a caller to optimize its
register usage by taking into account the actual registers clobbered at
a given call site.

Caller-saved registers are generally handling using a mask of preserved
registers as an operand to the call instruction. This pass manipulates
that mask to 'resurrect' registers, marking them as preserved if it can
be determined that the call target does not clobber that physical
register.

In order to determine the physical registers used at a call-site. The
callee must have completed register allocation, and should contain no
outstanding virtual registers. This means that most (currently all) of
the backend stage must be run in bottom-up call graph order.

The pass which implements the optimization is called RegisterResurrect,
which is a CallGraphSCCPass. It is not generally possible to run a
CallGraphSCCPass alongside MachineFunctionPasses because
MachineFunctionAnalysis becomes invalidated, so this also introduces a
fix which allows MachineFunctionAnalysis to be rerun, by caching the
Function -> MachineFunction mapping in MachineModuleInfo.

Currently, the physical registers used by a MachineFunction are
calculated once and then cached in MachineModuleInfo. This information
is probably available in MachineRegisterInfo anyway, so it should not
need to be recalculated.

There are a number of open issues with the current approach:

* To determine whether an instruction is a call, we use
  MachineInstr::isCall(), however there is no obvious mechanism to
  determine which of the operands is the target of the call.

* It is assumed that any register mask operand is the preserved register
  mask, and we therefore happily modify this mask with impunity.

* There is no way to stop a register from being resurrected, to force it
  to be caller saved.

The pass is enabled by default, and disabled with
-mllvm -disable-register-resurrect.