AAP: Update fixup and relaxation handling
Remove unnecessary code which was not being used. The backend does not
need to do any relaxation.

Add logic to apply a few generic fixups in the backend. This is
necessary for some values which need to be resolved by the time they
reach the linker (for example values held in debug information). AAP
specific fixups are still only resolved in the linker.