AAP: Handle short instructions more elegantly
Previously, we were selecting short instructions by adding AddedComplexity
until the short instructions were taking in preference. This has been
changed so that long instructions are always selected, but may be
shorted by a peephole pass which runs late on in the compile. The
peephole pass is called AAPShortInstrPeephole.cpp.

The peephole pass obviously can't handle assembly, so we still need the
operand classes to define their SuperClasses so that they are selected
in the prefered shorter-first order.

The handling of operand ranges has been refactored so that the main test
for valid operand values is now in AAP.h. AAPInstrInfo.td,
AsmParser/AAPAsmParser.cpp, and AAPShortInstrPeephole.cpp now all
reference these central test routines.