AAP: Allow immediates to contain relocatable expressions
Previously, immediate fields could only contain expressions which could
be evaluated by the compiler. Now they can contain anything, as long as
a relocation can be emitted for the expression.

This was always intended, but a mistake in the AsmParser meant that the
fields would only accept constants or absolute expressions.