AAP: Set a total order for memsrc3 and memsrc10 operands.
Previously, short MemSrc operands (MemSrc3, MemSrc3PostInc,
MemSrc3PreDec) were subclasses of their respectively long MemSrc
versions (MemSrc10, etc). This was not sufficient to ensure that the
assembly matcher tables were fully ordered.

The generation of the matcher tables relies first on the mnemonic, and
then on the operand classes. Since the mnemonics for all loads and
stores are the same (including postinc/predec), comparison would be
through operand classes only. When sorting the matcher table, we would
compare dissimilar classes (MemSrc3 with MemSrc10PostInc for example),
and in this case no order was specified and the table would only be
partially ordered.