AAP: Make var naming consistent, update getBranchOpForCondition
getBranchOpForCondition used to take the input operation as an argument,
because the returned branch opcode depended on whether the input branch
was a long or short instruction. This no longer matters, as we always
emit a long branch instruction.

The getBranchOpForCondition code was duplicated in emitSelectCC, and
this has been factored out.