AAP: Remove registers used for return from register mask operand
Previously this was acheived with a custom change, however
now there is now an upstream mechanism which has the same effect.

With AAP we have some registers which are callee-saved, but which
may also be used in the return. This change means that if a register
is used as a return, then it will not be callee saved.