AAP: Fix failure to select BRIND (indirect branch)
Added a pattern for the indirect branch node so it can be selected as a jmp
instruction. This change also uncovered that marking jmp as isReturn = 1 causes
issues as in some cases LLVM attempts to transform code using this information
even when the jmp is not used as a return.