AAP: Add basic support for the RETURNADDR DAG node
The RETURNADDR node is now lowered to a basic copy of the contents of the link
register when the operand is 0. Non-zero operands are currently unsupported
since there is no frame pointer register in AAP which makes traversing the stack
much more difficult. In the future the FRAMEADDR node should also be implemented
when traversing the stack is figured out.