AAP: Rewrite analyzeBranch & correct JMP behaviour
Since JMP was allowed to be generated as an indirect branch, the codegen has
struggled to correctly analyze it. This was because analyzeBranch did not expect
to deal with a JMP instruction, which was previously only used as a return. This
commit corrects the analyisis while also rearranging the function so that
the control flow makes more sense in this case. The instruction PseudoRET was
also added, as now JMP cannot be assumed as a return.