AAP: Correct erroneous instruction checks in tests
With the correct pseudo instruction expansion for PseudoRET, the
expanded form is written to the assembly file under -asm-show-inst, as
opposed to the string representation of the instruction which the tests
previously checked for.

The representation of the pseudo instruction should also be more
generic with regards to its inputs.