[AAP] Add codegen support for dynamic stack operations
Adds support for dynamic stack allocate, save and restore operations,
and adds support for using the frame pointer to restore the stack
pointer in function prologue & epilogues. To allow correct spill and
restore operation, frame index elimination was also implemented such
that the stack pointer is used for spill/restore even when the frame
pointer is available.